+++ app/src/main/java/com/fibelatti/photowidget/di/PhotoWidgetEntryPoint.kt
@@ -3,6 +3,7 @@ package com.fibelatti.photowidget.di
 import com.fibelatti.photowidget.configure.PhotoWidgetPinningCache
 import com.fibelatti.photowidget.configure.SavePhotoWidgetUseCase
 import com.fibelatti.photowidget.hints.HintStorage
+import com.fibelatti.photowidget.platform.ExceptionReporter
 import com.fibelatti.photowidget.platform.PhotoDecoder
 import com.fibelatti.photowidget.preferences.UserPreferencesStorage
 import com.fibelatti.photowidget.widget.CyclePhotoUseCase
@@ -40,4 +41,6 @@ interface PhotoWidgetEntryPoint {
     fun photoDecoder(): PhotoDecoder
 
     fun coroutineScope(): CoroutineScope
+
+    fun exceptionReporter(): ExceptionReporter
 }