+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Collect crash reports locally for optional reporting
+
 ## [v1.32.3] - 2025-09-07
 [v1.32.3]: https://github.com/fibelatti/photo-widget/releases/tag/v1.32.3